requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
tortoise-orm==0.17.8
aiosqlite==0.17.0
pydantic==1.8.2
websockets==10.0 